scrapping/pyproject.toml
[project]
name = "scrapping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "scrapy>=2.13.0",
    "scrapy-playwright>=0.0.43",
]
